MSI-576: Low Quantity Notification refactoring
