+++ app/code/Magento/InventoryLowQuantityNotification/Controller/Adminhtml/Report/ExportLowstockCsv.php
@@ -4,6 +4,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
+declare(strict_types=1);
 
 namespace Magento\InventoryLowQuantityNotification\Controller\Adminhtml\Report;
